Changed the file structure and fixed the webpack-dev-server hot reloading
