--- index.html
@@ -1,22 +1,0 @@
-<!DOCTYPE html>
-<html lang='en'>
-
-    <head>
-        <meta charset='UTF-8'>
-        <meta name='viewport' content='width=device-width, initial-scale=1.0'>
-        <meta http-equiv='X-UA-Compatible' content='ie=edge'>
-        <link rel='stylesheet' href='/src/styles/main.css'>
-        <title>Accenture AR</title>
-    </head>
-
-    <body>
-        <video id="video" width="auto" height="auto" playsinline hidden preload="auto"
-            poster="http://i.imgur.com/aWTM6Bk.gif">
-            <source src="/src/assets/videos/Tech-Vision-2019-Overview.mp4" type="video/mp4">
-            Your browser does not support HTML5 video tags.
-        </video>
-        <!-- include main.js file -->
-        <script src="dist/main.js"></script>
-    </body>
-
-</html>
