+++ webpack.config.js
@@ -1,11 +1,9 @@
 module.exports = {
-    target: 'node', // in order to ignore built-in modules like path, fs, etc.
-    watch: true,
-    devServer: {
-        index: 'index.html',
-        open: true
-        // watchOptions: {
-        //     poll: true
-        // }
+    entry: './src/js/main.js',
+    output: {
+        filename: './js/bundle.js'
+    },
+    node: {
+        fs: 'empty'
     }
 }
